% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif_adj_nondiff.R
\name{cif_adj_nondiff}
\alias{cif_adj_nondiff}
\title{A function to account for nondifferential misclassification in cumulative incidence functions}
\usage{
cif_adj_nondiff(data = data, data_evd = data_evd, t = "x",
  obs = "c_obs", true = "c", level = 1, tau = 1)
}
\arguments{
\item{data}{input data source containing main study data}

\item{data_evd}{intput data source containing validation data with variables true outcome and misclassified outcome}

\item{t}{name of the time variable}

\item{obs}{name of variable denoting possibly misclassified outcome type in both main and validation datasets}

\item{true}{name of variable denoting true outcome type (in validation data only)}

\item{level}{level of outcome for which to estimate cumulative incidence}

\item{tau}{maximum follow-up time}
}
\value{
data frame with event times and corrected cumulative incidence function for outcome of interest
}
\description{
A function to account for nondifferential misclassification in cumulative incidence functions
}
\examples{
cif_adj_nondiff(data = mydata, data_evd = valdata, t = "x", obs = "c_obs", true = "c", level = 1, tau = 2)
}
