% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif_naive.R
\name{cif_naive}
\alias{cif_naive}
\title{A function to estimate a nonparametric cumulative incidence function}
\usage{
cif_naive(data = data, t = "x", type = "c_obs", level = 1, tau = 1)
}
\arguments{
\item{data}{input data source containing main study data}

\item{t}{name of the time variable}

\item{type}{name of variable denoting outcome type}

\item{level}{level of outcome for which to estimate cumulative incidence}

\item{tau}{final follow-up time}
}
\value{
data frame with event times and cumulative incidence function for outcome of interest
}
\description{
A function to estimate a nonparametric cumulative incidence function
}
\examples{
cif_naive(mydata, t = "x", type = "c_obs", level = 1, tau = 2)
}
