% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassification_function.R
\name{lagpad}
\alias{lagpad}
\title{A helper function to account for differential misclassification}
\usage{
lagpad(x, k)
}
\arguments{
\item{x}{thing to lab}

\item{k}{how much to lag it}
}
\value{
vector
}
\description{
A helper function to account for differential misclassification
}
\examples{
lagpad(dsx$times,1)
}

