% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccr.R
\name{mccr}
\alias{mccr}
\title{Calculate the Matthews correlation coefficient (MCC) score}
\usage{
mccr(act, pred)
}
\arguments{
\item{act}{actual values (vector), 1 (positive), or 0 (negative)}

\item{pred}{predict values (vector), 1 (positive), or 0 (negative)}
}
\value{
MCC score
}
\description{
Calculate the Matthews correlation coefficient (MCC) score
}
\examples{
set.seed(18)
act <- abs(round(rnorm(100))) \%\% 2
pred <- abs(round(rnorm(100))) \%\% 2
mccr(act, pred)
}
