% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassification_function.R
\name{misclassification}
\alias{misclassification}
\title{A function to estimate misclassification parameters separately for each level of a binary covariate.}
\usage{
misclassification(data, c, c_obs, cov)
}
\arguments{
\item{data}{input data source containing external validation data}

\item{c}{variable name containing true outcome type}

\item{c_obs}{variable name containing misclassfied outcome type}

\item{diffx}{variable name containing covariate by which misclassification may be differential}
}
\value{
values of a and b  used in subsequent steps
}
\description{
A function to estimate misclassification parameters separately for each level of a binary covariate.
}
\examples{
misclassification(data, c, c_obs, diffx)
}
\seealso{
\code{\link{nchar}} which this function wraps
}

