% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassification_function.R
\name{misclassificationnd}
\alias{misclassificationnd}
\title{A function to estimate misclassification parameters overall}
\usage{
misclassificationnd(data, c, c_obs)
}
\arguments{
\item{data}{input data source containing external validation data}

\item{c}{variable name containing true outcome type}

\item{c_obs}{variable name containing misclassfied outcome type}
}
\value{
values of a and b  used in subsequent steps
}
\description{
A function to estimate misclassification parameters overall
}
\examples{
misclassificationnd(data, y, y_obs)
}

