% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif_adj.R
\name{cif_adj}
\alias{cif_adj}
\title{A function to account for differential misclassification in cumulative incidence functions}
\usage{
cif_adj(data, data_evd, tau, diff = 1)
}
\arguments{
\item{data}{input data source containing main study data, specifically time variable x and possibly misclassified outcome c_obs}

\item{data_evd}{intput data source containing validation data with variables true outcome c and misclassified outcome c_obs as well as any variable by which misclassification may be differential. This variable should be named diffx.}

\item{tau}{maximum follow-up time}

\item{diff}{indicator of whether misclassification is differential (diff=1) or nondifferential (diff=0)}
}
\value{
data frame with event times and corrected cumulative incidence function for outcome of interest
}
\description{
A function to account for differential misclassification in cumulative incidence functions
}
\examples{
cif_adj(mydata, valdata, 2, diff=1)
}

