% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif_adj_nondiff.R
\name{cif_adj_nondiff}
\alias{cif_adj_nondiff}
\title{A function to account for nondifferential misclassification in cumulative incidence functions}
\usage{
cif_adj_nondiff(data, data_evd, tau)
}
\arguments{
\item{data}{input data source containing main study data, specifically time variable x and possibly misclassified outcome c_obs}

\item{data_evd}{intput data source containing validation data}

\item{tau}{maximum follow-up time}
}
\value{
data frame with event times and corrected cumulative incidence function for outcome of interest
}
\description{
A function to account for nondifferential misclassification in cumulative incidence functions
}
\examples{
cif_adj_nondiff(mydata, valdata, 2)
}

