% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif_naive.R
\name{cif_naive}
\alias{cif_naive}
\title{A function to estimate a nonparametric cumulative incidence function}
\usage{
cif_naive(data, c, tau)
}
\arguments{
\item{data}{input data source containing main study data, specifically time variable x and outcome type c}

\item{c}{level of outcome for which to estimate cumulative incidence}

\item{tau}{final follow-up time}
}
\value{
data frame with event times and cumulative incidence function for outcome of interest
}
\description{
A function to estimate a nonparametric cumulative incidence function
}
\examples{
cif_naive(mydata, 1, 2)
}

